Fixed query for PR reviewers removing. Moved selectors to separate file. Fixed elements ids
