+++ src/client/containers/PullRequestSummary/ReviewSection/ReviewerList/index.js
@@ -43,10 +43,10 @@ class ReviewerList extends Component {
               />
             </OverlayTrigger>
             {this.props.reviews.map((r, i) => (
-              <span key={_.uniqueId()}>
+              <span key={_.uniqueId(r.user.username)}>
                 <span className="reviewer-item" onClick={() => this.handleRemoveReviewer(r.user.id)}>
-                  <a >
-                    {r.user.fullName.trim() || r.user.username.trim()}
+                  <a>
+                    {r.user.fullName.trim() || r.user.username.trim() || 'no name'}
                   </a>
                   <i className="fa fa-times-circle reviewer-remove-icon" aria-hidden="true"></i>
                 </span>
